# util/deephash: avoid using sync.Pool for reflect.MapIter